Move WindowIcons to IconChanger
