+++ src/main/java/meteordevelopment/meteorclient/systems/config/Config.java
@@ -18,7 +18,6 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.nbt.NbtString;
-import net.minecraft.util.Identifier;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -73,10 +72,11 @@ public class Config extends System<Config> {
         .build()
     );
 
-    public final Setting<WindowIcons> customWindowIcon = sgVisual.add(new EnumSetting.Builder<WindowIcons>()
+    @SuppressWarnings("unused")
+    public final Setting<IconChanger.WindowIcons> customWindowIcon = sgVisual.add(new EnumSetting.Builder<IconChanger.WindowIcons>()
         .name("custom-window-icon")
         .description("The icon to use for the window.")
-        .defaultValue(WindowIcons.Default)
+        .defaultValue(IconChanger.WindowIcons.Default)
         .onChanged(IconChanger::setIcon)
         .build()
     );
@@ -208,29 +208,5 @@ public class Config extends System<Config> {
         List<String> list = new ArrayList<>();
         for (NbtElement item : tag.getListOrEmpty(key)) list.add(item.asString().orElse(""));
         return list;
-    }
-
-    public enum WindowIcons {
-        Meteor(
-            MeteorClient.identifier("textures/icons/windowicon/meteor_16.png"),
-            MeteorClient.identifier("textures/icons/windowicon/meteor_32.png")
-        ),
-        Christmas(
-            MeteorClient.identifier("textures/icons/windowicon/christmas_16.png"),
-            MeteorClient.identifier("textures/icons/windowicon/christmas_32.png")
-        ),
-        Halloween(
-            MeteorClient.identifier("textures/icons/windowicon/halloween_16.png"),
-            MeteorClient.identifier("textures/icons/windowicon/halloween_32.png")
-        ),
-        Default(null, null);
-
-        public final Identifier icon16;
-        public final Identifier icon32;
-
-        WindowIcons(Identifier icon16, Identifier icon32) {
-            this.icon16 = icon16;
-            this.icon32 = icon32;
-        }
     }
 }
